+++ lib/pages/tag_page/tag_page.dart
@@ -1,4 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:piggram_mobile/components/picture_miniature_list.dart';
+import 'package:piggram_mobile/pages/tag_page/bloc/tag_page_bloc.dart';
 
 class TagPage extends StatelessWidget {
   final String tag;
@@ -10,9 +13,19 @@ class TagPage extends StatelessWidget {
       appBar: AppBar(
         title: Text("#$tag"),
       ),
-      body: Center(
-        child: Text(tag),
-      ),
+      body: BlocBuilder<TagPageBloc, TagPageState>(builder: (context, state) {
+        if (state is TagPageLoading) {
+          return Center(
+            child: CircularProgressIndicator(),
+          );
+        }
+        if (state is TagPageLoaded) {
+          return PictureMiniatureList(posts: state.posts);
+        }
+        return Center(
+          child: Text("Something went wrong, try it again"),
+        );
+      }),
     );
   }
 }
